implement storing pcap and graph files to cache/directory and move it to spool directory right after the call ends. the option is -C, --cachedir <dir> (cachedir in voipmonitor.conf)
store pcap and graph file to <cache/dir> and move it after call ends to spool directory. Moving all files are guaranteed to be serialized which 
solves slow random write I/O on magnetic or other media. Typical cache directory is /dev/shm/voipmonitor which is in RAM and grows
automatically or /mnt/ssd/voipmonitor which is mounted to SSD disk or some very fast SAS/SATA disk where spool can be network storage
or raid5 etc.

wav files are not implemented yet


git-svn-id: http://svn.code.sf.net/p/voipmonitor/code/trunk@378 937ffe21-195b-4c76-ad14-8d9f9b87382c