Now the voipmonitor returns memory back to linux and not grow giantly in case of call peaks.
You may encounter that voipmonitor process does not have a reduced memory usage although you freed the calls. 
This is because it allocates memory in a number of small chunks. When freeing one of those chunks, the OS may decide 
that giving this little memory back to the kernel will cause too much overhead and delay the operation. As all chunks 
are this small, they get actually freed but not returned to the kernel. On systems using glibc, there is a function call 
"malloc_trim" from malloc.h which does this missing operation (note that it is allowed to fail). 



git-svn-id: http://svn.code.sf.net/p/voipmonitor/code/trunk@616 937ffe21-195b-4c76-ad14-8d9f9b87382c